Map: вынес connect в отдельный модуль
