+++ src/components/map/index.js
@@ -1,0 +1,3 @@
+import Component from './map.connect';
+
+export default Component;
